🐛 Notes doesn't like ampersands in a palette's filename
When you import a palette with an ampersand in its filename (like
'Travel & Places.zip') Notes won't throw an error but after a restart
it just removes the palette.
To accomodate that, all ampersands will be replaced by the word 'and'
and spaces will be replaced by underscores (just for good measure).